attempts to clean up soe of the code climate errors.
