+++ eg/WinApi/locale_patches/win_api.py
@@ -1,0 +1,80 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of EventGhost.
+# Copyright © 2005-2019 EventGhost Project <http://www.eventghost.net/>
+#
+# EventGhost is free software: you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free
+# Software Foundation, either version 2 of the License, or (at your option)
+# any later version.
+#
+# EventGhost is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+# more details.
+#
+# You should have received a copy of the GNU General Public License along
+# with EventGhost. If not, see <http://www.gnu.org/licenses/>.
+
+
+import ctypes
+from ctypes.wintypes import DWORD, INT, WCHAR, LPCWSTR
+
+
+LOCALE_NAME_MAX_LENGTH = 85
+LOCALE_ALLOW_NEUTRAL_NAMES = 0x08000000
+LOCALE_SENGLISHLANGUAGENAME = 0x00001001
+LOCALE_SENGLISHCOUNTRYNAME = 0x00001002
+LOCALE_IDEFAULTANSICODEPAGE = 0x00001004
+LOCALE_SNATIVELANGUAGENAME = 0x00000004
+LOCALE_SLOCALIZEDLANGUAGENAME = 0x0000006f
+LOCALE_SNATIVECOUNTRYNAME = 0x00000008
+LOCALE_SLOCALIZEDCOUNTRYNAME = 0x00000006
+
+_kernel32 = ctypes.windll.Kernel32
+
+_GetLocaleInfoEx = _kernel32.GetLocaleInfoEx
+_GetLocaleInfoEx.restype = INT
+
+LCTYPE = DWORD
+
+
+def GetLocaleInfoEx(lp_locale_name, lc_type):
+    if not isinstance(lc_type, LCTYPE):
+        lc_type = LCTYPE(lc_type)
+
+    lp_lc_data = (WCHAR * 0)()
+
+    cch_data = _GetLocaleInfoEx(
+        LPCWSTR(lp_locale_name),
+        lc_type,
+        lp_lc_data,
+        0
+    )
+    if cch_data == 0:
+        return ''
+
+    lp_lc_data = (ctypes.c_wchar * cch_data)()
+    res = _GetLocaleInfoEx(
+        LPCWSTR(lp_locale_name),
+        lc_type,
+        lp_lc_data,
+        cch_data
+    )
+
+    if res == 0:
+        raise ctypes.WinError()
+
+    output = ''
+    for i in range(res):
+        char = lp_lc_data[i]
+        if char in ('\x00', 0x0):
+            break
+
+        output += char
+
+    try:
+        return int(output)
+    except ValueError:
+        return output
+
